Remove unused dependencies in backend
There were two unused dependencies in the node.js backend:
    * cookie-parser
    * body-parser

'body-parser' is not needed for Express v4.16.0, and 'cookie-parser'
is not needed since req.cookies does not need to be called.

Minor cleanup was also done on views/index.js to make it consistent with
the rest of the project.